requirements.txt
flask==2.0.1
flask-login==0.5.0
opencv-python==4.5.3.56
ultralytics==8.0.0
werkzeug==2.0.1
pillow==9.5.0 